+++ demo/demo/src/main/java/com/example/demo/model/vo/GoodInInfo.java
@@ -15,7 +15,7 @@ public class GoodInInfo {
 
     private String goodName;
 
-    private int goodPrice;
+    private float goodPrice;
 
     private String providerName;
