+++ demo/demo/src/main/java/com/example/demo/model/vo/GoodSaleInfo.java
@@ -11,11 +11,11 @@ public class GoodSaleInfo {
 
     int num;
 
-    int oneprice;
+    float oneprice;
 
-    int totalprice;
+    float totalprice;
 
-    int inprice;
+    float inprice;
 
     String goodname;
